+++ javaTypesGenerator/src/main/java/com/aws/jverify/generator/CSharpField.java
@@ -1,0 +1,14 @@
+package com.aws.jverify.generator;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CSharpField {
+    CSharpType type;
+    String name;
+
+    public CSharpField(CSharpType type, String name) {
+        this.type = type;
+        this.name = name;
+    }
+}
